Move error models back to their own package
