+++ src/main/java/ca/gov/dtsstn/passport/api/web/ElectronicServiceRequestController.java
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 import ca.gov.dtsstn.passport.api.service.NotificationService;
 import ca.gov.dtsstn.passport.api.service.PassportStatusService;
 import ca.gov.dtsstn.passport.api.web.exception.NonUniqueResourceException;
-import ca.gov.dtsstn.passport.api.web.model.BadRequestErrorModel;
 import ca.gov.dtsstn.passport.api.web.model.ElectronicServiceRequestModel;
-import ca.gov.dtsstn.passport.api.web.model.InternalServerErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.BadRequestErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.InternalServerErrorModel;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
