+++ src/main/java/ca/gov/dtsstn/passport/api/web/PassportStatusController.java
@@ -40,14 +40,14 @@ import ca.gov.dtsstn.passport.api.service.PassportStatusService;
 import ca.gov.dtsstn.passport.api.web.annotation.Authorities;
 import ca.gov.dtsstn.passport.api.web.exception.NonUniqueResourceException;
 import ca.gov.dtsstn.passport.api.web.exception.ResourceNotFoundException;
-import ca.gov.dtsstn.passport.api.web.model.AccessDeniedErrorModel;
-import ca.gov.dtsstn.passport.api.web.model.AuthenticationErrorModel;
-import ca.gov.dtsstn.passport.api.web.model.BadRequestErrorModel;
-import ca.gov.dtsstn.passport.api.web.model.InternalServerErrorModel;
 import ca.gov.dtsstn.passport.api.web.model.PassportStatusModel;
 import ca.gov.dtsstn.passport.api.web.model.PassportStatusModelAssembler;
-import ca.gov.dtsstn.passport.api.web.model.ResourceNotFoundErrorModel;
-import ca.gov.dtsstn.passport.api.web.model.UnprocessableEntityErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.AccessDeniedErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.AuthenticationErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.BadRequestErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.InternalServerErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.ResourceNotFoundErrorModel;
+import ca.gov.dtsstn.passport.api.web.model.error.UnprocessableEntityErrorModel;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
