+++ src/main/java/ca/gov/dtsstn/passport/api/web/model/error/AuthenticationErrorModel.java
@@ -1,44 +1,45 @@
-package ca.gov.dtsstn.passport.api.web.model;
+package ca.gov.dtsstn.passport.api.web.model.error;
 
 import java.io.Serializable;
 import java.time.Instant;
 
 import org.immutables.value.Value.Default;
 import org.immutables.value.Value.Immutable;
+import org.immutables.value.Value.Parameter;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.format.annotation.DateTimeFormat.ISO;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
 /**
- * REST model representing an HTTP 404 Not Found response.
+ * REST model representing an HTTP 401 Unauthorized response
  *
- * @author Sébastien Comeau (sebastien.comeau@hrsdc-rhdcc.gc.ca)
  * @author Greg Baker (gregory.j.baker@hrsdc-rhdcc.gc.ca)
  */
 @Immutable
-@Schema(name = "ResourceNotFoundError")
-public interface ResourceNotFoundErrorModel extends Serializable {
+@Schema(name = "AuthenticationError")
+public interface AuthenticationErrorModel extends Serializable {
 
 	@Default
-	@Schema(example = "404")
+	@Schema(example = "401")
 	default int getStatusCode() {
-		return 404;
+		return 401;
 	}
 
-	@Schema(example = "The resource with ID=[54543ab1-01b3-4edb-aad6-9b6c6b9e6985] was not found or the user does not have access")
+	@Parameter
+	@Schema(example = "Unauthorized.")
 	String getDetails();
 
 	@Default
-	@Schema(example = "API-0404")
+	@Schema(example = "API-0401")
 	default String getErrorCode() {
-		return "API-0404";
+		return "API-0401";
 	}
 
 	@Default
-	@Schema(example = "Not found")
+	@Schema(example = "Unauthorized")
 	default String getMessage() {
-		return "Not found";
+		return "Unauthorized";
 	}
 
 	@Default
